84 featurepostcontrollerservice 게시글 싫어요 시리즈 api

+++ src/like/like.service.ts
@@ -1,10 +1,12 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CommentLike } from './entities/comment-like.entity';
 import { Repository } from 'typeorm';
 import { PostLike } from './entities/post-like.entity';
 import { Post } from 'src/post/entities/post.entity';
 import { Comment } from 'src/comment/entities/comment.entity';
+import { COMMENT_MESSAGE } from 'src/constants/comment-message.constant';
+import { POST_MESSAGE } from 'src/constants/post-message.constant';
 
 @Injectable()
 export class LikeService {
@@ -17,37 +19,48 @@ export class LikeService {
     private readonly postLikeRepository: Repository<PostLike>,
     @InjectRepository(CommentLike)
     private readonly commentLikeRepository: Repository<CommentLike>
-  ) {}
+  ) { }
 
+  // 댓글 좋아요 조회 API
   async getCommentLikes(commentId: number) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
     }
 
+    // 댓글 좋아요 카운트
     const count = await this.commentLikeRepository.countBy({ commentId });
 
     return count;
   }
 
+  // 댓글 좋아요 생성 API
   async createCommentLike(userId, commentId) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+    const commentLike = await this.commentLikeRepository.findOneBy({
+      userId: userId,
+      commentId: commentId,
+    });
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
-    } else if ((existingComment.userId = userId)) {
-      throw new BadRequestException('자신의 댓글에 좋아요를 누를 수 없습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
     }
 
-    const commentLike = await this.commentLikeRepository.findOneBy({
-      userId,
-      commentId,
-    });
+    // 댓글 작성자 ID 비교 후 동일 유저이면 좋아요 금지
+    else if ((existingComment.userId === userId)) {
+      throw new BadRequestException(COMMENT_MESSAGE.LIKE.CREATE.FAILURE.NO_SELF);
+    }
+
+    // 댓글 좋아요 눌렀었는지 확인
     if (commentLike) {
-      throw new BadRequestException('이미 좋아요를 누른 댓글입니다.');
+      throw new BadRequestException(COMMENT_MESSAGE.LIKE.CREATE.FAILURE.ALREADY);
     }
 
     await this.commentLikeRepository.save({
@@ -56,12 +69,24 @@ export class LikeService {
     });
   }
 
+  // 댓글 좋아요 삭제 API
   async deleteCommentLike(userId, commentId) {
     const existingComment = await this.commentRepository.findOneBy({
       id: commentId,
     });
+    const commentLike = await this.commentLikeRepository.findOneBy({
+      userId: userId,
+      commentId: commentId,
+    })
+
+    // 댓글 존재 확인
     if (!existingComment) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+
+    // 내 좋아요 남아있는지 확인
+    if (!commentLike) {
+      throw new NotFoundException(COMMENT_MESSAGE.LIKE.DELETE.FAILURE.NO_LIKE);
     }
 
     await this.commentLikeRepository.delete({
@@ -70,12 +95,15 @@ export class LikeService {
     });
   }
 
+  // 게시글 좋아요 조회 API
   async getPostLikes(postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
+
+    // 게시글 존재 확인
     if (!existingPost) {
-      throw new BadRequestException('게시글이 존재하지 않습니다.');
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
     }
 
     const count = await this.postLikeRepository.countBy({ postId });
@@ -83,24 +111,30 @@ export class LikeService {
     return count;
   }
 
+  // 게시글 좋아요 생성 API
   async createPostLike(userId, postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
-    if (!existingPost) {
-      throw new BadRequestException('게시글이 존재하지 않습니다.');
-    } else if (existingPost.userId == userId) {
-      throw new BadRequestException(
-        '자신의 게시글에 좋아요를 누를 수 없습니다.'
-      );
-    }
 
     const postLike = await this.postLikeRepository.findOneBy({
       userId,
       postId,
     });
+
+    // 게시글 존재 확인
+    if (!existingPost) {
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
+    }
+
+    // 게시글 작성자 ID 비교 후 동일 유저이면 좋아요 금지
+    else if (existingPost.userId == userId) {
+      throw new BadRequestException(POST_MESSAGE.LIKE.CREATE.FAILURE.NO_SELF);
+    }
+
+    // 게시글 좋아요 눌렀었는지 확인
     if (postLike) {
-      throw new BadRequestException('이미 좋아요를 누른 댓글입니다.');
+      throw new BadRequestException(POST_MESSAGE.LIKE.CREATE.FAILURE.ALREADY);
     }
 
     await this.postLikeRepository.save({
@@ -109,12 +143,24 @@ export class LikeService {
     });
   }
 
+  // 게시글 좋아요 삭제 API
   async deletePostLike(userId, postId) {
     const existingPost = await this.postRepository.findOneBy({
       id: postId,
     });
+    const postLike = await this.postLikeRepository.findOneBy({
+      userId,
+      postId,
+    });
+
+    // 게시글 존재 확인
     if (!existingPost) {
-      throw new BadRequestException('댓글이 존재하지 않습니다.');
+      throw new NotFoundException(POST_MESSAGE.POST.NOT_FOUND);
+    }
+
+    // 내 좋아요 남아있는지 확인
+    if (!postLike) {
+      throw new NotFoundException(POST_MESSAGE.LIKE.DELETE.FAILURE.NO_LIKE);
     }
 
     await this.postLikeRepository.delete({
